Dogecoin: allow built-in miners to mine multiple block standards
Because Dogecoin does not allow AuxPoW on every block, not even
on regtest, we need to have a built-in miner that can mine
different block formats based on the nAuxPowVersion consensus
parameter.

This commit introduces a wrapper object CVersionedAuxBlock that
returns the correct headers to mine, mined PoW to compare,
adaptive setting/incrementing of nonces and whether or not to
initialize AuxPoW on a block.