Avoid redefining symbol on FreeBSD in scrypt.h
This commit unifies src/crypto/scrypt.h with the version in the 1.21-dev
branch and updates all calls appropriately.

Note that while the header guard defined in this file has changed from
`SCRYPT_H` to `BITCOIN_CRYPTO_SCRYPT_H`, the former does not appear in
the codebase on this branch at all, so that change should also be safe.

Note also that not all changes from src/crypto/scrypt.cpp in 1.21-dev
were merged into this file; only those related to this specific change
for FreeBSD.