fees: clarify what the actual recommended fee is
Introduces DEVELOPER_RECOMMENDED_FEE as a static value that
explicitly sets a single value and then implements this value
throughout the code where this is relevant, instead of using values
relative to COIN. This helps because it makes the relation of other
defaults to the fee recommendation clear and makes choices made by
developers in wallet configuration easier to follow.

This commit only changes existing values to be expressed relative
to the recommended fee and does not change actual values.